Refactoring duplicated code for credentials resources
* creating `VerifiedAllocatedResourceForm` to include all the common methods
* forms can extend this and implement the required methods
* refactoring GKE and EKS credentials forms to use this
