+++ ui/lib/components/forms/GKECredentialsForm.js
@@ -1,114 +1,15 @@
-import * as React from 'react'
-import PropTypes from 'prop-types'
-import canonical from '../../utils/canonical'
 import V1alpha1GKECredentials from '../../kore-api/model/V1alpha1GKECredentials'
 import V1alpha1GKECredentialsSpec from '../../kore-api/model/V1alpha1GKECredentialsSpec'
-import V1Allocation from '../../kore-api/model/V1Allocation'
-import V1AllocationSpec from '../../kore-api/model/V1AllocationSpec'
 import V1ObjectMeta from '../../kore-api/model/V1ObjectMeta'
-import V1Ownership from '../../kore-api/model/V1Ownership'
+import VerifiedAllocatedResourceForm from '../../components/forms/VerifiedAllocatedResourceForm'
 import ResourceVerificationStatus from '../../components/ResourceVerificationStatus'
+import FormErrorMessage from '../../components/forms/FormErrorMessage'
 import KoreApi from '../../kore-api'
-import { Button, Form, Input, Alert, Select, message, Typography, Card } from 'antd'
-const { Paragraph, Text } = Typography
+import { Button, Form, Input, Alert, Select, Card } from 'antd'
 
-class GKECredentialsForm extends React.Component {
-  static propTypes = {
-    form: PropTypes.any.isRequired,
-    team: PropTypes.string.isRequired,
-    allTeams: PropTypes.object,
-    data: PropTypes.object,
-    handleSubmit: PropTypes.func.isRequired,
-    saveButtonText: PropTypes.string,
-    inlineVerification: PropTypes.bool
-  }
+class GKECredentialsForm extends VerifiedAllocatedResourceForm {
 
-  constructor(props) {
-    super(props)
-    let allocations = []
-    if (props.data && props.data.allocation) {
-      allocations = props.data.allocation.spec.teams.filter(a => a !== '*')
-    }
-    this.state = {
-      submitting: false,
-      formErrorMessage: false,
-      allocations,
-      inlineVerificationFailed: false
-    }
-  }
-
-  componentDidMount() {
-    // To disabled submit button at the beginning.
-    this.props.form.validateFields()
-  }
-
-  disableButton = fieldsError => {
-    if (this.state.submitting) {
-      return true
-    }
-    return Object.keys(fieldsError).some(field => fieldsError[field])
-  }
-
-  onAllocationsChange = value => {
-    this.setState({
-      ...this.state,
-      allocations: value
-    })
-  }
-
-  async verify(gkeCredentials, tryCount) {
-    const messageKey = 'verify'
-    tryCount = tryCount || 0
-    if (tryCount === 0) {
-      message.loading({ content: 'Verifying GCP project credentials', key: messageKey, duration: 0 })
-    }
-    if (tryCount === 3) {
-      message.error({ content: 'GCP project credentials verification failed', key: messageKey })
-      this.setState({
-        ...this.state,
-        inlineVerificationFailed: true,
-        submitting: false,
-        formErrorMessage: (
-          <>
-            <Paragraph>The credentials have been saved but could not be verified, see the error below. Please try again or click &quot;Continue without verification&quot;.</Paragraph>
-            {(gkeCredentials.status.conditions || []).map((c, idx) =>
-              <Paragraph key={idx} style={{ marginBottom: '0' }}>
-                <Text strong>{c.message}</Text>
-                <br/>
-                <Text>{c.detail}</Text>
-              </Paragraph>
-            )}
-          </>
-        )
-      })
-    } else {
-      setTimeout(async () => {
-        const api = await KoreApi.client()
-        const gkeCredentialsResult = await api.GetGKECredential(this.props.team, gkeCredentials.metadata.name)
-        gkeCredentialsResult.allocation = gkeCredentials.allocation
-        if (gkeCredentialsResult.status.status === 'Success') {
-          message.success({ content: 'GCP project credentials verification successful', key: messageKey })
-          return await this.props.handleSubmit(gkeCredentialsResult)
-        }
-        return await this.verify(gkeCredentialsResult, tryCount + 1)
-      }, 2000)
-    }
-  }
-
-  setFormSubmitting = (submitting = true, formErrorMessage = false) => {
-    this.setState({
-      ...this.state,
-      submitting,
-      formErrorMessage
-    })
-  }
-
-  getMetadataName = values => {
-    const data = this.props.data
-    return (data && data.metadata && data.metadata.name) || canonical(values.name)
-  }
-
-  getGKECredentialsResource = values => {
+  generateGKECredentialsResource = values => {
     const resource = new V1alpha1GKECredentials()
     resource.setApiVersion('gke.compute.kore.appvia.io/v1alpha1')
     resource.setKind('GKECredentials')
@@ -127,75 +28,20 @@ class GKECredentialsForm extends React.Component {
     return resource
   }
 
-  getAllocationResource = values => {
+  getResource = async metadataName => {
+    const api = await KoreApi.client()
+    const gkeCredentialsResult = await api.GetGKECredential(this.props.team, metadataName)
+    gkeCredentialsResult.allocation = await api.GetAllocation(this.props.team, metadataName)
+    return gkeCredentialsResult
+  }
+
+  putResource = async values => {
+    const api = await KoreApi.client()
     const metadataName = this.getMetadataName(values)
-
-    const resource = new V1Allocation()
-    resource.setApiVersion('config.kore.appvia.io/v1')
-    resource.setKind('Allocation')
-
-    const meta = new V1ObjectMeta()
-    meta.setName(metadataName)
-    meta.setNamespace(this.props.team)
-    resource.setMetadata(meta)
-
-    const spec = new V1AllocationSpec()
-    spec.setName(values.name)
-    spec.setSummary(values.summary)
-    spec.setTeams(this.state.allocations.length > 0 ? this.state.allocations : ['*'])
-    const owner = new V1Ownership()
-    owner.setGroup('config.kore.appvia.io')
-    owner.setVersion('v1')
-    owner.setKind('GKECredentials')
-    owner.setName(metadataName)
-    owner.setNamespace(this.props.team)
-    spec.setResource(owner)
-
-    resource.setSpec(spec)
-
-    return resource
-  }
-
-  handleSubmit = e => {
-    e.preventDefault()
-
-    this.setFormSubmitting()
-
-    return this.props.form.validateFields(async (err, values) => {
-      if (err) {
-        this.setFormSubmitting(false, 'Validation failed')
-        return
-      }
-
-      const metadataName = this.getMetadataName(values)
-      try {
-        const api = await KoreApi.client()
-        const gkeResult = await api.UpdateGKECredential(this.props.team, metadataName, this.getGKECredentialsResource(values))
-        gkeResult.allocation = await api.UpdateAllocation(this.props.team, metadataName, this.getAllocationResource(values))
-
-        if (this.props.inlineVerification) {
-          await this.verify(gkeResult)
-        } else {
-          await this.props.handleSubmit(gkeResult)
-        }
-      } catch (err) {
-        console.error('Error submitting form', err)
-        this.setFormSubmitting(false, 'An error occurred saving the project credentials, please try again')
-      }
-    })
-  }
-
-  continueWithoutVerification = async () => {
-    try {
-      const metadataName = this.getMetadataName(this.props.form.getFieldsValue())
-      const api = await KoreApi.client()
-      const gkeCredentialsResult = await api.GetGKECredential(this.props.team, metadataName)
-      gkeCredentialsResult.allocation = await api.GetAllocation(this.props.team, metadataName)
-      await this.props.handleSubmit(gkeCredentialsResult)
-    } catch (err) {
-      console.error('Error getting data', err)
-      this.setFormSubmitting(false, 'An error occurred saving the project credentials, please try again')
-    }
+    const gkeResult = await api.UpdateGKECredential(this.props.team, metadataName, this.generateGKECredentialsResource(values))
+    const allocationResource = this.generateAllocationResource({ group: 'config.kore.appvia.io', version: 'v1', kind: 'GKECredentials' }, values)
+    gkeResult.allocation = await api.UpdateAllocation(this.props.team, metadataName, allocationResource)
+    return gkeResult
   }
 
   render() {
@@ -222,21 +68,6 @@ class GKECredentialsForm extends React.Component {
     const fieldError = fieldKey => isFieldTouched(fieldKey) && getFieldError(fieldKey)
     const allocationMissing = Boolean(data && !data.allocation)
 
-    const FormErrorMessage = () => {
-      if (formErrorMessage) {
-        return (
-          <Alert
-            message={formErrorMessage}
-            type="error"
-            showIcon
-            closable
-            style={{ marginBottom: '20px' }}
-          />
-        )
-      }
-      return null
-    }
-
     return (
       <div>
 
@@ -253,7 +84,7 @@ class GKECredentialsForm extends React.Component {
         ) : null}
 
         <Form {...formConfig} onSubmit={this.handleSubmit}>
-          <FormErrorMessage />
+          <FormErrorMessage message={formErrorMessage} />
           <Card style={{ marginBottom: '20px' }}>
             <Alert
               message="Project and service account"
